Bump start address one byte
The reason: I can't make the target function begin at the start of the
section. The reason is that a symbol, "$t", is always put in first.
According to some arm docs, this is to indicate that a section of thumb
code starts here, which isn't intensely useful on a thumb-only uC.
These are termed "mapping symbols", apparently.

I can't find a way to disable them, so plough onwards with some mildly
unaligned unpleasentness