Restructure flasher a little, eliminate DFUse
DFuse appears to be some kind of ST extention to DFU. It appears to
orient around some super specialised on-the-wire format, and the
host program reading / parsing the address space of the IC. While
interesting, it croaked (unexpected intr) for me repeatedly. Disable
this by
  a) removing a string descr containing it's data
  b) reset the DFU version to 1 (I think), not DFuse's version
  c) Remove the DFuse handling from the actual flashing function,
     usbdfu_getstatus_complete

Then hardcode the base address to be written. This could be bounds
checked better.

Finally, actually erase pages before writing them. I do not understand
how this ever worked before that. Sheesh.

This is an ugly patch, but I've been mangling this for a while, and this
is the version that actually shows a chance of working