Code-load USB DFU state variable
It gets written to from .data, and we can't trust that as it's outside
the bootloader.

I've looked at having two segments of .data to load, one from the
bootloader, and another from the rest of the program. It's do-able in LD
scripts, however it'd require patching libopencm3 as it only copies one
block of memory.