Revert incorrect attempt at bootloader munging
I'd got the incorrect impression that the bootloader was supposed to be
compiled in as part of the base firmware image. That isn't necessarily
wrong, but clearly they're supposed to be almost 100% decoupled. The
current arrangement, where parts of libopencm3 are in the bootloader,
will break when any firmware uses a new function because it'll be linked
into the bootloader portion of the image, but the bootloader portion of
the image will never be reflashed.

At the same time the state machine here should be available in both the
bootloader and application.

So instead I'm going to aim for generating an 8k blob that becomes an
elf file, but also an incrementally linked .o that's stripped and can be
installed at the start of any higher level application.