Don't produce relocatable bootloader
My understanding was that -r was just for "Link an output that can be
fed into LD again", reading further thought that precisely corresponds
to -r, 'Keep relocations'

This is undesireable, as the bootloader blob then has internal symbols
(such as 'main') left relocatable (i.e., not concrete). Stripping then
makes it impossible to correctly link the output.

So remove -r option. LD still seems to be able to eat the output, and
now the bootloaders reset_handler actually initializes sram and calls
the bootloader main.