Add a program for computing a crc32 of a file
Significantly, existing tools appear to work on a byte-level basis. This
leads to some inaccuracy: the stm32 hardware feeds bits into it's crc
unit the way an ethernet frame would be, while things like crc32(1) and
online tools deal only in bytes. The result is mismatching crc values.

To get around this, use the boost crc implementation, configured to be
the relevant (ethernet) crc. While this adds a dependency, it also
raises the fewest licensing questions.