Add a suitable DFU interface descriptor string
This string is read and parsed by host DFU programming software in order to
determine some properties of the layout of the STM32's memory space. For
example, dfu-programmer [1] uses this information to find the address and size
of the flash segment, and which sub-region within it is occupied by the
bootloader.

See User Manual UM0424 section 4.3.2 for a specification of the format. This
string describes:
- 8 1KiB sectors starting at 0x08000000, read-only (the DFU bootloader)
- 24 1KiB sectors starting at 0x08002000, read-write (the application segment)

[1] http://git.infradead.org/users/borneoa/dfu-programmer.git